README: Clarify to avoid using the clasp rootDir option
We are bundling and pushing a single dist file, so
that we can use es6 imports in all our files.
Given that, the .claspingnore file is super simple
(mentions just 2 files), which allows us to avoid
having to use the rootDir .clasp.json option and also
copy more files to the dist folder before pushing.

Change-type: patch